pyproject.toml
[tool.poetry]
name = "redactify"
version = "0.1.0"
description = "Clean text files from sensitive information"
authors = ["Michael Segal <michaelseg14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spacy = "^3.8.2"
fpdf = "^1.7.2"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
